refactor: modularize sorter
